+++ EnhanceQoL/Modules/Aura/Init.lua
@@ -19,6 +19,8 @@ function addon.Aura.functions.InitDB()
 	init("resourceBarsHideOutOfCombat", false)
 	init("resourceBarsHideMounted", false)
 	init("resourceBarsHideVehicle", false)
+	init("resourceBarsHidePetBattle", false)
+	if addon.db.resourceBarsHidePetBattle == nil and addon.db.auraHideInPetBattle ~= nil then addon.db.resourceBarsHidePetBattle = addon.db.auraHideInPetBattle and true or false end
 
 	-- spec specific settings for personal resource bars
 	init("personalResourceBarSettings", {})
@@ -54,5 +56,4 @@ function addon.Aura.functions.InitDB()
 	init("cooldownPanelsEditorY", 0)
 
 	if addon.Aura and addon.Aura.CooldownPanels and addon.Aura.CooldownPanels.NormalizeAll then addon.Aura.CooldownPanels:NormalizeAll() end
-
 end